+++ bumpmap.rt
@@ -1,0 +1,4 @@
+A  0.4      100,100,200
+L  3,0,-5  0.8         255,0,0
+C  0,0,-10 0,0,1 90
+co 0,0,0 -1,0,0 90 120,120,200
