+++ srcs/rt_object_cone.c
@@ -6,30 +6,12 @@
 /*   By: corvvs <corvvs@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/08 20:13:53 by corvvs            #+#    #+#             */
-/*   Updated: 2021/12/26 20:56:22 by corvvs           ###   ########.fr       */
+/*   Updated: 2021/12/26 23:38:14 by corvvs           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minirt.h"
 #define EPS 1e-9
-
-void	rt_set_tangent_cone(
-	t_hit_record *rec
-)
-{
-	const t_vec3 pc = mr_vec3_sub(rec->p, rec->element.position);
-
-	rec->normal = mr_vec3_sub(
-		mr_vec3_mul_double(&pc, cos(rec->element.fov * M_PI / 360) / mr_vec3_length(&pc)),
-		rec->element.direction
-	);
-	rec->normal = mr_unit_vector(&rec->normal);
-	rec->w0 = rec->normal;
-	rec->v0 = mr_unit_vector(&pc);
-	rec->u0 = mr_vec3_cross(&rec->v0, &rec->w0);
-	t_vec3 vt = {0, 0, 1};
-	rec->normal = rt_vec_tangent_to_global(rec, &vt);
-}
 
 static bool	t_predicate(
 	const t_element *el,
@@ -75,20 +57,6 @@ static bool actual_hittest(
 	return (false);
 }
 
-void	rt_texture_cone(t_hit_record *rec, const t_element *el)
-{
-	const t_vec3	r = mr_vec3_sub(rec->p, el->position);
-	const double	phi = mr_vec3_dot(el->direction, r);
-	const t_vec3	u1 = rt_coord_perpendicular_unit(&el->direction);
-	const t_vec3	u2 = rt_coord_turn_around_90(&u1, &el->direction);
-	const double	dx = mr_vec3_dot(r, u1);
-	const double	dz = mr_vec3_dot(r, u2);
-	const double	theta = atan2(dz, dx);
-
-	rec->tex.u = theta / M_PI;
-	rec->tex.v = phi / (2 * M_PI);
-}
-
 bool	rt_hittest_cone(
 	const t_element *el,
 	const t_ray *ray,
@@ -96,6 +64,6 @@ bool	rt_hittest_cone(
 {
 	if (!actual_hittest(el, ray, rec))
 		return (false);
-	rt_texture_cone(rec, el);
+	// rt_texture_cone(rec, el);
 	return (true);
 }
