+++ srcs/rt_object_plane.c
@@ -6,7 +6,7 @@
 /*   By: corvvs <corvvs@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/08 20:14:54 by corvvs            #+#    #+#             */
-/*   Updated: 2021/12/26 20:56:29 by corvvs           ###   ########.fr       */
+/*   Updated: 2021/12/26 23:30:52 by corvvs           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,18 +30,6 @@ t_vec3	rt_vec_tangent_to_global(
 		+ rec->v0.z * vtangent->y
 		+ rec->w0.z * vtangent->z;
 	return (vglobal);
-}
-
-void	rt_set_tangent_plane(
-	t_hit_record *rec
-)
-{
-	rec->normal = rec->element.direction;
-	rec->w0 = rec->element.direction;
-	rec->u0 = rt_coord_perpendicular_unit(&rec->w0);
-	rec->v0 = rt_coord_turn_around_90(&rec->u0, &rec->w0);
-	t_vec3 vt = {0, 0, 1};
-	rec->normal = rt_vec_tangent_to_global(rec, &vt);
 }
 
 bool	rt_hittest_plane(
